ft_intlen long

+++ ft_intlen.c
@@ -6,13 +6,13 @@
 /*   By: djeannot <djeannot@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/27 16:13:40 by djeannot          #+#    #+#             */
-/*   Updated: 2022/04/27 16:28:38 by djeannot         ###   ########.fr       */
+/*   Updated: 2022/04/28 10:17:32 by djeannot         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_intlen(int n)
+int	ft_intlen(long n)
 {
 	int	i;
