+++ libft.h
@@ -6,7 +6,7 @@
 /*   By: djeannot <djeannot@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/04 10:17:48 by djeannot          #+#    #+#             */
-/*   Updated: 2022/04/27 16:16:29 by djeannot         ###   ########.fr       */
+/*   Updated: 2022/04/28 10:17:51 by djeannot         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -73,6 +73,6 @@ void	ft_lstiter(t_list *lst, void (*f)(void *));
 
 // Added function
 
-int		ft_intlen(int n);
+int		ft_intlen(long n);
 
 #endif
